% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_isotops.R
\name{graphic_df}
\alias{graphic_df}
\title{Graphical function}
\usage{
graphic_df(
  df,
  img.title. = img.title,
  iso.left. = iso.left,
  iso.right. = iso.right,
  y.left.color. = y.left.color,
  y.right.color. = y.right.color,
  y.left.by. = y.left.by,
  y.right.by. = y.right.by,
  x.by. = x.by,
  stat.ci. = stat.ci,
  min.left. = min.left,
  min.right. = min.right,
  max.left. = max.left,
  max.right. = max.right,
  xlim.ERJ. = xlim.ERJ,
  a. = a,
  b. = b,
  flag. = flag
)
}
\description{
Graphical function
}
