% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_ca.R
\name{zoo_ca}
\alias{zoo_ca}
\title{Correspondance Analysis of the NISP data}
\usage{
zoo_ca(
  df = NA,
  num_column = "num",
  site_column = "site",
  period_column = "period",
  percBOTA_column = "percBOTA",
  percOC_column = "percOC",
  percSUDO_column = "percSUDO",
  typSite_column = "type.site",
  lorder_period = NA,
  typsit_symb = NA,
  CA1_interval = c(-1, 2),
  CA2_interval = c(-1, 1),
  colVar = "black",
  shpVar = 17,
  pt_siz = 1.5
)
}
\arguments{
\item{df}{a dataframe}

\item{num_column}{the column name of assemblage numbers}

\item{site_column}{the column name of assemblage sites}

\item{period_column}{the column name of assemblage periods}

\item{percBOTA_column}{the column name of assemblage percents of BOTA}

\item{percOC_column}{the column name of assemblage percents of OC}

\item{percSUDO_column}{the column name of assemblage percents of SUDO}

\item{typSite_column}{the column name of assemblage site types}

\item{lorder_period}{a vector with the ordered periods}

\item{typsit_symb}{a dataframe with the symbology}

\item{CA1_interval}{the min and max intervals for the F1 axis}

\item{CA2_interval}{the min and max intervals for the F2 axis}

\item{colVar}{the color of the variables}

\item{shpVar}{the shape of the variables}
}
\value{
a gglot with the different CA by period
}
\description{
Correspondance Analysis of the NISP data by periods
}
\examples{

df <- zoo_read()
lorder_period <- zoo_order_period(df)
typsit_symb <- zoo_legends(worksheets = c("sites_types"))
zoo_ca(df = df, lorder_period = lorder_period, typsit_symb = typsit_symb)

}
