% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_isotops.R
\name{zoo_isotops}
\alias{zoo_isotops}
\title{Isotopic variation on two y-columns.}
\arguments{
\item{dataDir}{directory of the dataset}

\item{dataFile}{name of the dataset}

\item{img.title}{title of the output plot}

\item{iso.names}{names of the isotopes.By default columns c("O", "Sr")}

\item{iso.colors}{color of the isotopes. By default c("turquoise", "turquoise4")}

\item{iso.by.y}{intervals on y-axis. By default c(5, .0002)}

\item{x.by}{interval on x-axis. By default = 5}

\item{stat.ci}{show statistics: confidence interval. Default value is TRUE}

\item{img.layout}{layout for the output. By default c("grid", "align")}

\item{img.dim}{dimensions for the output, Width and length. By default list(grid = c(18, 16), align = c(13, 21))}

\item{img.units}{By default "cm"}

\item{col.ERJ}{x-axis. By default column 'ERJ'}

\item{col.ind}{group column. By default column 'ind' (individus)
be removed?}
}
\value{
ggplot graphs, in grid layout or aligned vertically.
}
\description{
Compare two isotopic variations (eg O, Sr) on a two y-axis plots
}
\examples{
zoo_isotops()

zoo_isotops(img.layout = "align", 
            img.format = ".png",
            iso.by = c(1, .0002),
            x.by = 5)

zoo_isotops(img.layout = "grid",
            iso.colors = c("red", "blue"),
            stat.ci = FALSE,
            img.format = ".png")
}
