% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_legends.R
\name{zoo_legends}
\alias{zoo_legends}
\title{Legends for sites, taxons, etc.}
\usage{
zoo_legends(
  dataDir = paste0(system.file(package = "zoowork"), "/extdata/"),
  dataFile = "legends.xlsx",
  worksheets = c("sites_types", "sites_zones", "taxon_types"),
  outDir = paste0(system.file(package = "zoowork"), "/extdata/")
)
}
\arguments{
\item{dataDir}{directory of the legend workbook.
By default the package folder}

\item{dataFile}{name of the legend workbook}

\item{worksheets}{names of the different worksheets to be read 
By default: c("sites_types", "sites_zones", etc.)}

\item{outDir}{output folder
By default the package folder}
}
\value{
Series of plot showing the legends (shapes and colors)
}
\description{
Read a XLSX file where legends are recorded in different worksheets. 
Plot different legends for: sites' types, sites' zones (geographical areas), etc.
}
\examples{
zoo_legends()

zoo_legends(worksheets = c("sites_types", "sites_zones"))
}
