% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_order_period.R
\name{zoo_order_period}
\alias{zoo_order_period}
\title{Read a XLSX dataset}
\usage{
zoo_order_period(
  df = NA,
  period_column = "period",
  lorder_period = c("EN", "MN", "LN", "EBA", "MBA", "LBA", "EIA", "MIA1", "MIA2", "RR",
    "ERE", "LRE", "LA")
)
}
\arguments{
\item{df}{a dataframe}

\item{period_column}{the column name of assemblage periods}

\item{lorder_period}{a vector with the ordered periods}
}
\value{
the ordered list of periods existing in the dataframe
}
\description{
Test if the periods existing in the dataset are also in the order
reference period list. If not return an error
}
\examples{

df <- zoo_read()
lorder_period <- zoo_order_period(df)

}
