% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_read.R
\name{zoo_read}
\alias{zoo_read}
\title{Read a XLSX dataset}
\usage{
zoo_read(
  xlsx = paste0(getwd(),
    "/inst/extdata/Type_Site_PlosOne_CAT_14_TOM_rvTH_6_ZOOWORK.xlsx"),
  sheet = 1,
  num_column = "num",
  key_column = "key",
  period_column = "period"
)
}
\arguments{
\item{xlsx}{the complete path of the XLSX dataset, by default the test one}

\item{sheet}{the index or the name of the XLSX sheet to read, by default the
first one}

\item{num_column}{the column name of assemblage numbers}

\item{key_column}{the column name of assemblage keys}

\item{period_column}{the column name of assemblage periods}
}
\value{
A dataframe
}
\description{
Read a XLSX dataset
}
\examples{

df <- zoo_read()

}
